[BE] refactor: dto 분리
- issue (#169)
