+++ server/src/domains/group/dto/ApplyGroupDto.ts
@@ -6,8 +6,3 @@ export class ApplyGroupDto {
   @IsNumber()
   userId: number;
 }
-
-export class ApplyParam {
-  @IsNumber()
-  aid: number;
-}
