+++ server/src/domains/group/dto/ApplyParam.ts
@@ -1,0 +1,6 @@
+import { IsNumber } from "class-validator";
+
+export class ApplyParam {
+  @IsNumber()
+  aid: number;
+}
