+++ server/src/domains/group/dto/updateGroup/UpdateGroupDateDto.ts
@@ -1,0 +1,10 @@
+import { IsDateString, IsNumber } from "class-validator";
+
+export class UpdateGroupDateDto {
+  @IsNumber()
+  gid: number;
+  @IsDateString()
+  startAt: string;
+  @IsDateString()
+  endAt: string;
+}
