+++ server/src/domains/group/dto/updateGroup/UpdateGroupIntroDto.ts
@@ -1,0 +1,9 @@
+import { IsNumber, IsString, Length } from "class-validator";
+
+export class UpdateGroupIntroDto {
+  @IsNumber()
+  gid: number;
+  @IsString()
+  @Length(1, 500)
+  intro: string;
+}
