+++ server/src/domains/group/dto/updateGroup/UpdateGroupNameDto.ts
@@ -1,0 +1,9 @@
+import { IsNumber, IsString, Length } from "class-validator";
+
+export class UpdateGroupNameDto {
+  @IsNumber()
+  gid: number;
+  @IsString()
+  @Length(1, 20)
+  name: string;
+}
