+++ server/src/domains/group/dto/updateGroup/index.ts
@@ -1,0 +1,3 @@
+export * from './UpdateGroupNameDto';
+export * from './UpdateGroupDateDto';
+export * from './UpdateGroupIntroDto';
